requirements.txt
attrs==23.2.0
Automat==22.10.0
cffi==1.16.0
constantly==23.10.4
cryptography==42.0.7
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
pycparser==2.22
pyOpenSSL==24.1.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
PyYAML==6.0.1
setuptools==70.0.0
six==1.16.0
tk==0.1.0
Twisted==24.3.0
typing_extensions==4.12.0
zope.interface==6.4.post2
